+++ tests/plugins/test_pyfai_integration_base.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -16,25 +18,25 @@
 """Unit tests for pydidas modules."""
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 
 
-import unittest
-import tempfile
-import shutil
-import os
 import logging
+import os
+import shutil
+import tempfile
+import unittest
 from pathlib import Path
 
-import pyFAI
 import numpy as np
+import pyFAI
 
-from pydidas.plugins import BasePlugin, pyFAIintegrationBase
-from pydidas.core import get_generic_parameter, UserConfigError
 from pydidas.contexts import DiffractionExperimentContext
+from pydidas.core import UserConfigError, get_generic_parameter
+from pydidas.plugins import BasePlugin, pyFAIintegrationBase
 
 
 EXP = DiffractionExperimentContext()
@@ -44,6 +46,16 @@ logger.setLevel(logging.ERROR)
 
 
 class TestPyFaiIntegrationBase(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):
+        cls._det_dist = 0.1
+        EXP.set_param_value("detector_dist", cls._det_dist)
+        cls._det_pxsize = 1000
+        EXP.set_param_value("detector_pxsizex", cls._det_pxsize)
+        EXP.set_param_value("detector_pxsizey", cls._det_pxsize)
+        cls._lambda = 0.1
+        EXP.set_param_value("xray_wavelength", cls._lambda)
+
     def setUp(self):
         self._temppath = tempfile.mkdtemp()
         self._shape = (50, 50)
@@ -97,6 +109,87 @@ class TestPyFaiIntegrationBase(unittest.TestCase):
         plugin = pyFAIintegrationBase(rad_unit="Q / nm^-1")
         self.assertEqual(plugin.get_pyFAI_unit_from_param("rad_unit"), "q_nm^-1")
 
+    def test_modulate_and_store_azi_range__rad_ranges_flipped(self):
+        plugin = pyFAIintegrationBase(
+            azi_use_range="Specify azimuthal range",
+            azi_unit="chi / rad",
+            azi_range_lower=0.2,
+            azi_range_upper=0.1,
+        )
+        with self.assertRaises(UserConfigError):
+            plugin.modulate_and_store_azi_range()
+
+    def test_modulate_and_store_azi_range__rad_mod_2pi(self):
+        plugin = pyFAIintegrationBase(
+            azi_use_range="Specify azimuthal range",
+            azi_unit="chi / rad",
+            azi_range_lower=0.1 - 2 * np.pi,
+            azi_range_upper=0.2 - 2 * np.pi,
+        )
+        plugin.modulate_and_store_azi_range()
+        _low, _high = plugin.get_azimuthal_range_native()
+        self.assertAlmostEqual(_low, 0.1)
+        self.assertAlmostEqual(_high, 0.2)
+
+    def test_modulate_and_store_azi_range__rad__low_lt_zero(self):
+        plugin = pyFAIintegrationBase(
+            azi_use_range="Specify azimuthal range",
+            azi_unit="chi / rad",
+            azi_range_lower=-0.4,
+            azi_range_upper=0.2,
+        )
+        plugin.modulate_and_store_azi_range()
+        _low, _high = plugin.get_azimuthal_range_native()
+        self.assertAlmostEqual(_low, -0.4)
+        self.assertAlmostEqual(_high, 0.2)
+
+    def test_modulate_and_store_azi_range__rad_mod_pi(self):
+        plugin = pyFAIintegrationBase(
+            azi_use_range="Specify azimuthal range",
+            azi_unit="chi / rad",
+            azi_range_lower=0.1 - np.pi,
+            azi_range_upper=0.2 - np.pi,
+        )
+        plugin.modulate_and_store_azi_range()
+        _low, _high = plugin.get_azimuthal_range_native()
+        self.assertAlmostEqual(_low, 0.1 - np.pi)
+        self.assertAlmostEqual(_high, 0.2 - np.pi)
+
+    def test_modulate_and_store_azi_range__deg_ranges_flipped(self):
+        plugin = pyFAIintegrationBase(
+            azi_use_range="Specify azimuthal range",
+            azi_unit="chi / deg",
+            azi_range_lower=0.2,
+            azi_range_upper=0.1,
+        )
+        with self.assertRaises(UserConfigError):
+            plugin.modulate_and_store_azi_range()
+
+    def test_modulate_and_store_azi_range__deg_mod_2pi(self):
+        plugin = pyFAIintegrationBase(
+            azi_use_range="Specify azimuthal range",
+            azi_unit="chi / deg",
+            azi_range_lower=0.1 - 360,
+            azi_range_upper=0.2 - 360,
+        )
+        plugin.modulate_and_store_azi_range()
+        _range = plugin.get_azimuthal_range_native()
+        _low, _high = plugin.get_azimuthal_range_native()
+        self.assertAlmostEqual(_low, 0.1)
+        self.assertAlmostEqual(_high, 0.2)
+
+    def test_modulate_and_store_azi_range__deg_mod_pi(self):
+        plugin = pyFAIintegrationBase(
+            azi_use_range="Specify azimuthal range",
+            azi_unit="chi / deg",
+            azi_range_lower=0.1 - 180,
+            azi_range_upper=0.2 - 180,
+        )
+        plugin.modulate_and_store_azi_range()
+        _low, _high = plugin.get_azimuthal_range_native()
+        self.assertAlmostEqual(_low, 0.1 - 180)
+        self.assertAlmostEqual(_high, 0.2 - 180)
+
     def test_get_azimuthal_range_native__no_range(self):
         plugin = pyFAIintegrationBase(azi_use_range="Full detector")
         _range = plugin.get_azimuthal_range_native()
@@ -110,15 +203,6 @@ class TestPyFaiIntegrationBase(unittest.TestCase):
         )
         _range = plugin.get_azimuthal_range_native()
         self.assertIsNone(_range)
-
-    def test_get_azimuthal_range_native__upper_bound_lower(self):
-        plugin = pyFAIintegrationBase(
-            azi_use_range="Specify azimuthal range",
-            azi_range_lower=15,
-            azi_range_upper=12,
-        )
-        _range = plugin.get_azimuthal_range_native()
-        self.assertEqual(_range, (12, 15))
 
     def test_get_azimuthal_range_native__correct(self):
         _range = (12, 37)
@@ -232,6 +316,51 @@ class TestPyFaiIntegrationBase(unittest.TestCase):
         )
         _newrange = plugin.get_radial_range()
         self.assertEqual(_range, _newrange)
+
+    def test_get_radial_range_as_2theta(self):
+        _input_low, _input_high = (12, 37)
+        plugin = pyFAIintegrationBase(
+            rad_use_range="Specify radial range",
+            rad_unit="2theta / deg",
+            rad_range_lower=_input_low,
+            rad_range_upper=_input_high,
+        )
+        _low, _high = plugin.get_radial_range_as_2theta()
+        self.assertAlmostEqual(_input_low, _low)
+        self.assertAlmostEqual(_input_high, _high)
+
+    def test_get_radial_range_as_r(self):
+        _input_low, _input_high = (12, 37)
+        plugin = pyFAIintegrationBase(
+            rad_use_range="Specify radial range",
+            rad_unit="2theta / deg",
+            rad_range_lower=_input_low,
+            rad_range_upper=_input_high,
+        )
+        _low, _high = plugin.get_radial_range_as_r()
+        _target_low = self._det_dist * np.tan(_input_low * np.pi / 180) * 1e3
+        _target_high = self._det_dist * np.tan(_input_high * np.pi / 180) * 1e3
+        self.assertAlmostEqual(_low, _target_low)
+        self.assertAlmostEqual(_high, _target_high)
+
+    def test_get_radial_range_as_q(self):
+        _input_low, _input_high = (12, 37)
+        plugin = pyFAIintegrationBase(
+            rad_use_range="Specify radial range",
+            rad_unit="2theta / deg",
+            rad_range_lower=_input_low,
+            rad_range_upper=_input_high,
+        )
+        _low, _high = plugin.get_radial_range_as_q()
+        # factor 10 is 1e-9 from q in nm times 1e10 from lambda in A
+        _target_low = (
+            4 * np.pi / self._lambda * np.sin(_input_low * np.pi / 180 / 2) * 10
+        )
+        _target_high = (
+            4 * np.pi / self._lambda * np.sin(_input_high * np.pi / 180 / 2) * 10
+        )
+        self.assertAlmostEqual(_low, _target_low)
+        self.assertAlmostEqual(_high, _target_high)
 
     def test_calculate_result_shape(self):
         _range = (1234, 789)